mejora relacion bodega y posicion
