+++ bodega/src/commands/posicion/crear_posicion.py
@@ -13,7 +13,7 @@ class CrearPosicion(BaseCommand):
     
     def check_campos_requeridos(self) -> bool:
 
-        required_fields = ['bodega', 'volumen']
+        required_fields = ['nombre_posicion', 'bodega', 'volumen']
 
         if not all(field in self.posicion_template for field in required_fields):
             return False
@@ -22,24 +22,35 @@ class CrearPosicion(BaseCommand):
             return False
 
         return True
+    
         
     def verificar_bodega_existe(self) -> bool:
         existe_bodega_query = Bodega.query.filter(
-            Bodega.id == self.posicion_template['bodega']
+            Bodega.nombre == self.posicion_template['bodega']
         ).first()
         if existe_bodega_query:
             return True
         else:
             return False
         
-    def agregar_posicion_a_bodega(self, id_posicion: str, id_bodega: str):
+    def obtener_id_bodega(self) -> str:
+        existe_bodega_query = Bodega.query.filter(
+            Bodega.nombre == self.posicion_template['bodega']
+        ).first()
+
+        if existe_bodega_query:
+            return existe_bodega_query.id
+        else:
+            return None
         
-        bodega = Bodega.query.filter(Bodega.id == id_bodega).first()
+    def agregar_posicion_a_bodega(self, nombre_posicion: str, bodega: str):
+        
+        bodega = Bodega.query.filter(Bodega.nombre == bodega).first()
 
         if not bodega.posiciones:
             bodega.posiciones = []
 
-        bodega.posiciones = bodega.posiciones + [id_posicion]
+        bodega.posiciones = bodega.posiciones + [nombre_posicion]
         db.session.commit()
 
     def execute(self):
@@ -60,16 +71,27 @@ class CrearPosicion(BaseCommand):
                 "status_code": 404
             }
         
+        id_bodega = self.obtener_id_bodega()
+        if not id_bodega:
+            return {
+                "response": {
+                    "msg": "Error al obtener el id de la bodega."
+                },
+                "status_code": 500
+            }
+        
         id_posicion = self.crear_uuid()
 
         nueva_posicion = Posicion(
             id=id_posicion,
+            nombre_posicion=self.posicion_template['nombre_posicion'],
             bodega=self.posicion_template['bodega'],
+            id_bodega=id_bodega,
             volumen=self.posicion_template['volumen']
         )
 
         try:
-            self.agregar_posicion_a_bodega(id_posicion, self.posicion_template['bodega'])
+            self.agregar_posicion_a_bodega(self.posicion_template['nombre_posicion'], self.posicion_template['bodega'])
         
         except Exception as e:
             return {
@@ -88,6 +110,7 @@ class CrearPosicion(BaseCommand):
                 "response": {
                     "posicion": {
                         "id": nueva_posicion.id,
+                        "nombre_posicion": nueva_posicion.nombre_posicion,
                         "bodega": nueva_posicion.bodega,
                         "volumen": nueva_posicion.volumen
                     },
@@ -101,7 +124,7 @@ class CrearPosicion(BaseCommand):
             db.session.rollback()
             return {
                 "response": {
-                    "msg": "Error al crear la posicion",
+                    "msg": f"Error al crear la posicion: {str(e)}"
                 },
                 "status_code": 500
             }
