+++ bodega/tests/commands/posicion/test_buscar_posicion.py
@@ -3,10 +3,11 @@ import json
 from faker import Faker
 
 from app import app
-from src.commands.posicion.crear_posicion import CrearPosicion
+from src.commands.posicion.buscar_posicion import BuscarPosicion
 from src.commands.base_command import BaseCommand
 
-class TestCrearPosicion():
+
+class TestBuscarPosicion():
 
     @pytest.fixture(scope='module')
     def gen_request_posicion(self):
@@ -15,6 +16,7 @@ class TestCrearPosicion():
 
         for i in range(10):
             request_body = {
+                'nombre_posicion': fake.name(),
                 'volumen': fake.pyfloat(),
             }
             request_bodies.append(request_body)
@@ -39,18 +41,21 @@ class TestCrearPosicion():
     
     def test_base_model_inherit(self, gen_request_posicion):
         # Test to ensure RouteCreation inherits from BaseCommand
-        route = CrearPosicion(gen_request_posicion[0])
+        route = BuscarPosicion({"posicion": "1"})
         assert isinstance(route, BaseCommand)
 
-    def test_crear_posicion(self, gen_request_posicion, gen_request_bodega):
+    def test_buscar_posicion(self, gen_request_posicion, gen_request_bodega):
         with app.test_client() as client:
 
             response_bodega = client.post('/crear_bodega', json=gen_request_bodega[0])
-            id_bodega = response_bodega.json['bodega']['id']
+            nombre_bodega = response_bodega.json['bodega']['nombre']
 
             request_body = gen_request_posicion[0]
-            request_body['bodega'] = id_bodega
+            request_body['bodega'] = nombre_bodega
 
             response_posicion = client.post('/crear_posicion', json=request_body)
-            assert response_posicion.status_code == 201
-            assert response_posicion.json["msg"] == "Posicion creada correctamente"
+
+            id_posicion = response_posicion.json['posicion']['id']
+
+            response_buscar_posicion = client.get('/buscar_posicion', json={"posicion": id_posicion})
+            assert response_buscar_posicion.status_code == 200
