+++ bodega/tests/commands/posicion/test_crear_posicion.py
@@ -3,11 +3,10 @@ import json
 from faker import Faker
 
 from app import app
-from src.commands.posicion.buscar_posicion import BuscarPosicion
+from src.commands.posicion.crear_posicion import CrearPosicion
 from src.commands.base_command import BaseCommand
 
-
-class TestBuscarPosicion():
+class TestCrearPosicion():
 
     @pytest.fixture(scope='module')
     def gen_request_posicion(self):
@@ -16,6 +15,7 @@ class TestBuscarPosicion():
 
         for i in range(10):
             request_body = {
+                'nombre_posicion': fake.name(),
                 'volumen': fake.pyfloat(),
             }
             request_bodies.append(request_body)
@@ -40,21 +40,18 @@ class TestBuscarPosicion():
     
     def test_base_model_inherit(self, gen_request_posicion):
         # Test to ensure RouteCreation inherits from BaseCommand
-        route = BuscarPosicion({"id_posicion": "1"})
+        route = CrearPosicion(gen_request_posicion[0])
         assert isinstance(route, BaseCommand)
 
-    def test_buscar_posicion(self, gen_request_posicion, gen_request_bodega):
+    def test_crear_posicion(self, gen_request_posicion, gen_request_bodega):
         with app.test_client() as client:
 
             response_bodega = client.post('/crear_bodega', json=gen_request_bodega[0])
-            id_bodega = response_bodega.json['bodega']['id']
+            nombre_bodega = response_bodega.json['bodega']['nombre']
 
             request_body = gen_request_posicion[0]
-            request_body['bodega'] = id_bodega
+            request_body['bodega'] = nombre_bodega
 
             response_posicion = client.post('/crear_posicion', json=request_body)
-
-            id_posicion = response_posicion.json['posicion']['id']
-
-            response_buscar_posicion = client.get('/buscar_posicion', json={"id_posicion": id_posicion})
-            assert response_buscar_posicion.status_code == 200
+            assert response_posicion.status_code == 201
+            assert response_posicion.json["msg"] == "Posicion creada correctamente"
