+++ bodega/tests/commands/posicion/test_listar_posicion.py
